refactor: Replace the wizard scroll gate with a shared "more below" hint
## Summary
- Drop the hard scroll-to-bottom gate from the creation wizard (follow-up
  decision): overflowing steps no longer disable Next/Create. The "more content
  below" cues become pure hints.
- Extract the cues — a down-chevron disc, a bottom fade gradient, and a one-time
  scroller flash — into a shared `ScrollMoreIndicator`, used by both the wizard
  steps and the delete-VM sheet so the affordance is identical across the app.

## Changes
- New `Views/Shared/ScrollMoreIndicator.swift`: geometry-driven (clip bounds +
  document frame) hint that fades the chevron/fade overlays in and out as content
  scrolls (no latch — it tracks position) and flashes the scroller once on
  overflow. Owns hit-transparent overlays inserted into the scroll view's
  superview; exposes `hasMoreBelow` for tests.
- Remove the gate: `VMCreationViewModel.currentStepScrollGateSatisfied` (+ setter,
  per-step reset, and the "scroll down" validation hint); the shell's button
  state reverts to `canAdvance`/`canCreate`, and the gate-only keyboard-scroll
  plumbing (`WizardRootView` / `WizardScrollKey`) and forced layout are gone.
- The four wizard step VCs and `DeleteVMSheetContentViewController` each create a
  `ScrollMoreIndicator`; the delete sheet's bespoke `viewDidAppear` scroller flash
  is replaced by it (its `contentOverflows` sizing flag stays). Delete
  `WizardScrollGate`; move the chevron/fade atoms out of `WizardStyle`.

## Test plan
- [x] `make build` / `make lint` clean
- [x] `make test` green; new `ScrollMoreIndicatorTests` (geometry + overlay
      insertion); gate-specific model/shell tests removed
- [x] Live: wizard overflowing steps show the chevron + fade with Next/Create
      enabled (no gate); the delete sheet uses the same shared cue

## Notes
- Issue #374 requested the hard gate; after seeing it, the decision was to keep
  hints only and drop the gate. This delivers the indicator, not the gate.
