+++ Kernova/Views/Creation/ReviewContentViewController.swift
@@ -12,9 +12,9 @@ import AppKit
 final class ReviewContentViewController: NSViewController {
     private let creationVM: VMCreationViewModel
     private let startSwitch = NSSwitch()
-    /// Gates Create until the user has scrolled this (often-overflowing) summary
-    /// to the bottom; reports scroll geometry into the shared model.
-    private var scrollGate: WizardScrollGate?
+    /// Shows the "more content below" cue while this (often-overflowing) summary
+    /// doesn't fit the sheet; a hint only.
+    private var scrollMoreIndicator: ScrollMoreIndicator?
 
     init(creationVM: VMCreationViewModel) {
         self.creationVM = creationVM
@@ -46,9 +46,7 @@ final class ReviewContentViewController: NSViewController {
         ])
 
         view = scrollView
-        scrollGate = WizardScrollGate(scrollView: scrollView) { [weak creationVM] satisfied in
-            creationVM?.setCurrentStepScrollGateSatisfied(satisfied)
-        }
+        scrollMoreIndicator = ScrollMoreIndicator(scrollView: scrollView)
     }
 
     private func makeSummary() -> NSView {
