+++ Kernova/Views/Creation/WizardStyle.swift
@@ -121,92 +121,6 @@ func makeWizardRadioOption(radio: NSButton, iconSymbol: String, description desc
     return option
 }
 
-// MARK: - Scroll indicator
-
-/// A passive overlay container that lets clicks and scroll-wheel events fall
-/// through to the content beneath.
-///
-/// Used for the "more below" scroll indicator so the chevron never intercepts the
-/// scrolling it's prompting.
-final class WizardHitTransparentView: NSView {
-    // RATIONALE: Returning nil from hitTest removes the whole subtree from mouse/
-    // scroll event routing, so events reach the scroll view layered beneath the
-    // overlay — the standard AppKit pattern for a non-interactive overlay.
-    override func hitTest(_ point: NSPoint) -> NSView? { nil }
-}
-
-/// Builds the "more content below — scroll to continue" indicator: a `chevron.down`
-/// on a grey disc, sized for the bottom-center of a scrolling wizard step.
-///
-/// The disc is an `NSBox` so its fill/border are `NSColor`s that adapt to light/dark
-/// automatically; the fill sits well above the grouped-form cards' faint tint so the
-/// disc reads against the form content, and a hairline border defines its edge. The
-/// returned view is hit-transparent (see ``WizardHitTransparentView``), so the caller
-/// can layer it over the step's scroll view without blocking scrolling. The caller
-/// owns positioning and show/hide (the wizard shell fades it based on
-/// `VMCreationViewModel.currentStepScrollGateSatisfied`).
-@MainActor
-func makeWizardScrollIndicator() -> NSView {
-    let diameter: CGFloat = 28
-
-    let disc = NSBox()
-    disc.boxType = .custom
-    disc.titlePosition = .noTitle
-    disc.cornerRadius = diameter / 2
-    disc.fillColor = .secondaryLabelColor.withAlphaComponent(0.2)
-    disc.borderWidth = 1
-    disc.borderColor = .separatorColor
-
-    let chevron = NSImageView(
-        image: .systemSymbol(
-            "chevron.down", accessibilityDescription: "More content below — scroll to continue"))
-    chevron.symbolConfiguration = NSImage.SymbolConfiguration(pointSize: 13, weight: .semibold)
-    chevron.contentTintColor = .secondaryLabelColor
-    chevron.translatesAutoresizingMaskIntoConstraints = false
-
-    // RATIONALE: a custom NSBox sizes a `contentView` through the legacy autoresizing
-    // path and collapses, so it's pinned as a chrome layer behind the chevron sibling
-    // (the same pattern as `makeGroupedFormBox`).
-    let container = WizardHitTransparentView()
-    container.translatesAutoresizingMaskIntoConstraints = false
-    container.addFullSizeSubview(disc)
-    container.addSubview(chevron)
-
-    NSLayoutConstraint.activate([
-        container.widthAnchor.constraint(equalToConstant: diameter),
-        container.heightAnchor.constraint(equalToConstant: diameter),
-        chevron.centerXAnchor.constraint(equalTo: container.centerXAnchor),
-        chevron.centerYAnchor.constraint(equalTo: container.centerYAnchor),
-    ])
-    return container
-}
-
-/// Height of the bottom fade strip (see ``WizardScrollFadeView``).
-let wizardScrollFadeHeight: CGFloat = 36
-
-/// A passive bottom-edge fade for a scrolling wizard step.
-///
-/// The content dissolves into the sheet background as it nears the bottom — a soft
-/// "more below" cue that complements the chevron. Drawn with an `NSGradient` in
-/// `draw(_:)` (re-evaluated under the current appearance, so it adapts to light/dark
-/// without `CGColor` juggling) and hit-transparent so it never blocks scrolling.
-final class WizardScrollFadeView: NSView {
-    override var isOpaque: Bool { false }
-
-    // RATIONALE: nil from hitTest drops the view from event routing so the scroll
-    // view beneath still scrolls under the cursor — same pattern as
-    // ``WizardHitTransparentView``.
-    override func hitTest(_ point: NSPoint) -> NSView? { nil }
-
-    override func draw(_ dirtyRect: NSRect) {
-        let opaque = NSColor.windowBackgroundColor
-        let clear = opaque.withAlphaComponent(0)
-        // 90° runs bottom→top: opaque background at the bottom edge fading to clear,
-        // so content shows through above and dissolves into the sheet at the bottom.
-        NSGradient(starting: opaque, ending: clear)?.draw(in: bounds, angle: 90)
-    }
-}
-
 // MARK: - Buttons & badges
 
 /// Builds a borderless, link-styled push button (caption font, link color).
